backend/requirements.txt
fastapi
uvicorn
pydantic
rich
pyserial

